+++ apps/site/src/app/design-system/[brand]/[...component]/components/content-tabs/components/responsive-modal-button/responsive-modal.component.tsx
@@ -10,8 +10,8 @@ export function ResponsiveModal({ children, ...props }: ResponsiveModalProps) {
 
   return (
     <Modal isDismissable {...props}>
-      <ModalBody className="flex flex-1 flex-col px-0">
-        <GridContainer fixed className="flex flex-1 flex-col pt-5" id="demo-content">
+      <ModalBody className="flex flex-1 flex-col px-0" id="demo-content">
+        <GridContainer fixed className="flex flex-1 flex-col pt-5">
           <h2 className="typography-body-7 font-black xsl:typography-body-6">
             Viewport: <span className="font-light">{width}px</span>
           </h2>