+++ packages/ui/src/components/date-picker-beta/components/date-field/components/date-segment/date-segment.styles.ts
@@ -2,18 +2,14 @@ import { tv } from 'tailwind-variants';
 
 export const styles = tv(
   {
-    base: 'font-light disabled:form-control-disabled focus:focus-outline',
+    base: 'uppercase disabled:form-control-disabled focus:focus-outline',
     variants: {
-      isPlaceholder: {
-        true: 'text-text-50 opacity-100',
-        false: '',
-      },
       isFocusVisible: {
         true: 'focus-outline',
         false: '',
       },
       isSeparator: {
-        true: 'px-0.5 text-text-50',
+        true: 'px-0.5 text-text',
         false: '',
       },
     },