+++ packages/ui/src/components/date-picker-beta/date-picker.component.tsx
@@ -26,6 +26,8 @@ export function DatePickerBeta({
   disableDaysOfWeek,
   disableWeekends,
   separator,
+  placement = 'bottom left',
+  block = false,
   portalContainer,
   ...props
 }: DatePickerBetaProps) {
@@ -47,7 +49,13 @@ export function DatePickerBeta({
   }, [disableDaysOfWeek, disableWeekends, isDateUnavailable, locale]);
 
   const state = useDatePickerState({ isDateUnavailable: enhancedIsDateUnavailable, ...props });
-  const styles = datePickerStyles({ size, isInvalid: state.isInvalid, isDisabled: props.isDisabled });
+  const styles = datePickerStyles({
+    size,
+    isInvalid: state.isInvalid,
+    isDisabled: props.isDisabled,
+    isReadOnly: props.isReadOnly,
+    block,
+  });
   const breakpoint = useBreakpoint();
   const ref = useRef(null);
 
@@ -84,8 +92,19 @@ export function DatePickerBeta({
   return (
     <>
       <div {...labelProps}>{props.label}</div>
-      <div {...props} {...groupProps} ref={ref} className={styles.input({ className })}>
-        <DateField separator={separator} {...fieldProps} />
+      <div
+        {...props}
+        {...groupProps}
+        ref={ref}
+        onBlur={e => {
+          if (state.value) {
+            return props.onBlur?.(e, state.value);
+          }
+          return props.onBlur?.(e);
+        }}
+        className={styles.input({ className })}
+      >
+        <DateField separator={separator} {...fieldProps} className={styles.dateField()} />
         <Button
           look="faint"
           className={styles.button()}
@@ -102,7 +121,7 @@ export function DatePickerBeta({
           showAsBottomSheet={showAsBottomSheet}
           state={state}
           triggerRef={ref}
-          placement="bottom left"
+          placement={placement}
         >
           <Dialog {...dialogProps}>
             <Calendar {...calendarProps} firstDayOfWeek={props.firstDayOfWeek} />