+++ packages/ui/src/components/date-picker-beta/date-picker.stories.tsx
@@ -1,6 +1,9 @@
 import { DateValue, getLocalTimeZone, isWeekend, today } from '@internationalized/date';
 import { type Meta, StoryFn, type StoryObj } from '@storybook/react-vite';
+import { useState } from 'react';
 import { useLocale } from 'react-aria';
+
+import { Field } from '../index.js';
 
 import { DatePickerBeta } from './date-picker.component.js';
 
@@ -103,6 +106,44 @@ export const Disabled: Story = {
 };
 
 /**
+ * > Controlled
+ */
+export const Controlled = () => {
+  const now = today(getLocalTimeZone());
+  const disabledRanges = [
+    [now, now.add({ days: 5 })],
+    [now.add({ days: 14 }), now.add({ days: 16 })],
+    [now.add({ days: 23 }), now.add({ days: 24 })],
+  ];
+
+  const { locale } = useLocale();
+  const [date, setDate] = useState<DateValue | null>(null);
+  const [blurredDate, setBlurredDate] = useState<DateValue | null>(null);
+  const isDateUnavailable = (date: DateValue) =>
+    isWeekend(date, locale) ||
+    disabledRanges.some(interval => date.compare(interval[0]) >= 0 && date.compare(interval[1]) <= 0);
+
+  return (
+    <>
+      <p>onChange Date: {date ? date.toString() : 'None'}</p>
+      <p>
+        onBlur Date (NOTE: Focus returns to button after using calendar so blur event not fired):{' '}
+        {blurredDate ? blurredDate.toString() : 'None'}
+      </p>
+      <DatePickerBeta
+        onBlur={(e, date) => {
+          if (date) setBlurredDate(date);
+        }}
+        onChange={e => {
+          setDate(e as DateValue);
+        }}
+        isDateUnavailable={isDateUnavailable}
+      />
+    </>
+  );
+};
+
+/**
  * > Disabled dates
  */
 export const DisabledDates: Story = {
@@ -192,5 +233,37 @@ export const ShowAsBottomSheetResponsive: Story = {
   args: {},
   render: () => {
     return <DatePickerBeta bottomSheetView={{ initial: false, lg: true }} />;
+  },
+};
+
+/**
+ * > Different sizes example
+ */
+export const DatePickerBlock: Story = {
+  args: { block: true },
+};
+
+/**
+ * > Placement of popover
+ */
+export const Placement: Story = {
+  args: {},
+  render: () => {
+    return (
+      <div className="mt-[200px] flex flex-col space-y-2">
+        <Field label="Bottom Left">
+          <DatePickerBeta placement="bottom left" block />
+        </Field>
+        <Field label="Bottom Right">
+          <DatePickerBeta placement="bottom right" block />
+        </Field>
+        <Field label="Top Left">
+          <DatePickerBeta placement="top left" block />
+        </Field>
+        <Field label="Top Right">
+          <DatePickerBeta placement="top right" block className="bottom-0" />
+        </Field>
+      </div>
+    );
   },
 };