+++ packages/ui/src/components/date-picker-beta/date-picker.styles.ts
@@ -3,38 +3,51 @@ import { tv } from 'tailwind-variants';
 export const styles = tv(
   {
     slots: {
-      input: 'form-control flex items-center border-borderDark disabled:form-control-disabled',
-      button:
-        'flex h-auto items-center justify-center rounded-l-none border-y-0 border-l border-r-0 border-l-borderDark bg-light',
+      input: 'form-control flex items-center border-0',
+      dateField: 'rounded rounded-r-none border border-r-0 border-borderDark',
+      button: 'flex items-center justify-center rounded-l-none border-borderDark bg-light',
     },
     variants: {
       size: {
         small: {
-          input: 'form-control-small gap-1.5',
-          button: '-my-0.5 -mr-1.5 mb-[-0.25rem] min-h-5 py-[0.25rem]',
+          input: 'max-h-5',
+          dateField: 'form-control-small h-5',
         },
         medium: {
-          input: 'form-control-medium gap-2',
-          button: 'my-[-0.3125rem] -mr-2 min-h-6 py-[0.3125rem]',
+          input: 'max-h-6',
+          dateField: 'form-control-medium h-6',
         },
         large: {
-          input: 'form-control-large gap-2.5',
-          button: 'my-[-0.5rem] -mr-2.5 min-h-7 py-[0.5rem]',
+          input: 'max-h-7',
+          dateField: 'form-control-large h-7',
         },
         xlarge: {
-          input: 'form-control-xlarge gap-3',
-          button: '-my-1.5 -mr-3 mb-[-0.625rem] min-h-8 py-1.5',
+          input: 'max-h-8',
+          dateField: 'form-control-xlarge h-8',
+        },
+      },
+      block: {
+        true: {
+          input: 'w-full gap-0',
+          dateField: 'w-full',
         },
       },
       isInvalid: {
         true: {
-          input: 'border-danger',
+          dateField: 'border-danger',
+          button: 'border-danger border-l-borderDark',
         },
         false: {},
       },
       isDisabled: {
         true: {
-          input: 'form-control-disabled',
+          dateField: 'form-control-disabled',
+        },
+        false: {},
+      },
+      isReadOnly: {
+        true: {
+          dateField: 'form-control-disabled',
         },
         false: {},
       },