+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,11 @@
 # playground
 
+## 0.1.2
+
+### Patch Changes
+
+- 925e9f0: package updates
+
 ## 0.1.1
 
 ### Patch Changes